backend/requirements.txt
# Core Framework
fastapi==0.123.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20
pydantic==2.12.5
pydantic-settings==2.7.1
email-validator==2.2.0
python-dotenv==1.0.1

# Document Processing
pdfplumber==0.11.8
python-docx==1.2.0
docxtpl==0.18.0
docx2pdf==0.1.8
pypdf==5.1.0

# AI/ML & LangChain
langchain==1.1.0
langchain-community==0.3.15
langchain-google-genai==3.2.0
google-generativeai==0.8.5
chromadb==1.3.5
sentence-transformers==5.1.2

# Google ADK (Agent Development Kit)
google-adk

# Database & Storage
sqlalchemy==2.0.36
alembic==1.14.0

# Utilities
python-dotenv==1.2.1
httpx==0.28.1
aiofiles==24.1.0
jinja2==3.1.5

# Logging & Monitoring
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Development
black==24.10.0
ruff==0.8.5
mypy==1.14.0

# Web UI
streamlit==1.40.2
